Fixed pull request
I changed the button size and margins on mobile and changed the function names to be more suggestive. Also I deleted the useless parts from ts(ngOnInit and constructors) as we don't use them yet. I made the cursor a hand on hover and made the bottom and right equal.